+++ ProjectSourceCode/Game/gamelogger.cpp
