+++ [COMP345] Assignment 1.vcxproj
@@ -128,8 +128,8 @@
     <ClCompile Include="ProjectSourceCode\Interactable.cpp" />
     <ClCompile Include="ProjectSourceCode\MapBuilder.cpp" />
     <ClCompile Include="ProjectSourceCode\Wall.cpp" />
-    <ClCompile Include="ProjectSourceCode\game.cpp" />
-    <ClCompile Include="ProjectSourceCode\gamelogger.cpp" />
+    <ClCompile Include="ProjectSourceCode\Game\game.cpp" />
+    <ClCompile Include="ProjectSourceCode\Game\gamelogger.cpp" />
     <ClCompile Include="ProjectSourceCode\testgamelogger.cpp" />
   </ItemGroup>
   <ItemGroup>
@@ -175,8 +175,8 @@
     <ClInclude Include="ProjectSourceCode\Interactable.h" />
     <ClInclude Include="ProjectSourceCode\MapBuilder.h" />
     <ClInclude Include="ProjectSourceCode\Wall.h" />
-    <ClInclude Include="ProjectSourceCode\game.h" />
-    <ClInclude Include="ProjectSourceCode\gamelogger.h" />
+    <ClInclude Include="ProjectSourceCode\Game\game.h" />
+    <ClInclude Include="ProjectSourceCode\Game\gamelogger.h" />
     <ClInclude Include="ProjectSourceCode\testgamelogger.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
